refactor: move 'useObservableEffect' to 'useObservable.ts'
