--- src/hooks/useObservableEffect.ts
@@ -1,9 +1,0 @@
-import { useEffect } from 'react'
-import { Observable } from 'rxjs'
-
-export function useObservableEffect<T>(factory: () => Observable<T>, effect: (x: T) => void, deps: unknown[] = []) {
-  useEffect(() => {
-    const sub = factory().subscribe(effect)
-    return () => sub.unsubscribe()
-  }, deps)
-}
